refactor: move CLI argument definitions to command submodules
- Move CLI argument structs from main monocle.rs to individual command modules
- Each command module now contains its own Args struct using clap::Args
- Reduces main file from ~1600 lines to ~140 lines
- Improves code organization and maintainability
- No functional changes to CLI behavior
